refactor(client): split api.service monolith into per-domain services (A17.2)
The 1187-line ApiService held ~120 HTTP-wrapper methods + ~450 lines of DTO
declarations and was injected by two dozen components. Split into:
  - api.types.ts — all 50 shared DTOs (import a type without a service)
  - eight per-domain services over inject(HttpClient): AuthApi, SpacesApi,
    SchemaApi, BrainApi, FilesApi, DuplicatesApi, NetworksApi, AdminApi

All 24 consumers (components + specs) now inject only the domain services they
use. Every endpoint URL, query param, and method signature is unchanged —
pure structural move. Verified by the Angular AOT production build (a
mis-routed call is a compile error since each service exposes only its own
methods) and the full client unit suite (71 tests pass).
