--- client/src/app/core/api.service.ts
@@ -1,1187 +1,0 @@
-import { Injectable, inject } from '@angular/core';
-import { HttpClient, HttpParams, HttpHeaders } from '@angular/common/http';
-import { Observable, Subscription } from 'rxjs';
-import { map } from 'rxjs/operators';
-
-// ── Shared types ─────────────────────────────────────────────────────────────
-
-export interface Space {
-  id: string;
-  label: string;
-  builtIn?: boolean;
-  folders?: string[];
-  maxGiB?: number;
-  usageGiB?: number;
-  description?: string;
-  proxyFor?: string[];
-  meta?: SpaceMeta;
-  dupeRules?: DupeActionRule[];
-  dupeMergeSurvivor?: 'older' | 'newer';
-  dupeRulesOnInsert?: boolean;
-  /** Vector-index build state for a newly created space (B1). 'building' while the
-   *  Atlas indexes finish; absent means ready. Semantic recall waits for READY. */
-  indexStatus?: 'building' | 'ready' | 'failed';
-  /** Networks this space belongs to (F8). Absent/empty when the space is in no
-   *  network — the Brain chip shows the network indicator only when present. */
-  networks?: { id: string; label: string; type: Network['type'] }[];
-  /** Aggregate sync/governance status across this space's networks (F8), for the
-   *  chip indicator. 'vote' = an open round affects it; 'degraded' = a peer has
-   *  failed repeatedly (investigate); 'syncing' = a cycle is running; 'idle' =
-   *  member, nothing active. There is no true "fully synced" state (eventual sync). */
-  networkStatus?: 'vote' | 'degraded' | 'syncing' | 'idle';
-}
-
-export type ValidationMode = 'off' | 'warn' | 'strict';
-export type KnowledgeType = 'entity' | 'memory' | 'edge' | 'chrono';
-
-export interface PropertySchema {
-  type?: 'string' | 'number' | 'boolean' | 'date';
-  enum?: (string | number | boolean)[];
-  minimum?: number;
-  maximum?: number;
-  pattern?: string;
-  mergeFn?: 'avg' | 'min' | 'max' | 'sum' | 'and' | 'or' | 'xor';
-  required?: boolean;
-  default?: string | number | boolean;
-}
-
-export interface TypeSchema {
-  /** Reference to a schema library entry. Format: `"library:<name>"`. */
-  $ref?: string;
-  namingPattern?: string;
-  tagSuggestions?: string[];
-  propertySchemas?: Record<string, PropertySchema>;
-}
-
-/** An entry in the instance-level schema library. */
-export interface SchemaLibraryEntry {
-  name: string;
-  knowledgeType: KnowledgeType;
-  typeName: string;
-  schema: Omit<TypeSchema, '$ref'>;
-  description?: string;
-  /** Optional group tag for organizing related entries into a named set. */
-  schemaGroup?: string;
-  published?: boolean;
-  sourceUrl?: string;
-  sourceCatalog?: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface SchemaCatalog {
-  name: string;
-  url: string;
-  description?: string;
-  createdAt: string;
-  /** True if an access token is stored server-side for this catalog (token itself is never returned). */
-  hasAccessToken?: boolean;
-}
-
-/** A single entry from a foreign catalog (public endpoint shape). */
-export interface ForeignCatalogEntry {
-  name: string;
-  knowledgeType: KnowledgeType;
-  typeName: string;
-  description?: string;
-  updatedAt: string;
-  /** Present on full entry fetch, absent on index listing. */
-  schema?: Omit<TypeSchema, '$ref'>;
-}
-
-export interface SpaceMeta {
-  version?: number;
-  purpose?: string;
-  usageNotes?: string;
-  validationMode?: ValidationMode;
-  typeSchemas?: Partial<Record<KnowledgeType, Record<string, TypeSchema>>>;
-  tagSuggestions?: string[];
-  strictLinkage?: boolean;
-  updatedAt?: string;
-}
-
-export interface SpaceMetaResponse extends SpaceMeta {
-  spaceId: string;
-  spaceName: string;
-  stats: SpaceStats;
-}
-
-export interface SpacesResponse {
-  spaces: Space[];
-  storage?: {
-    usageGiB: { files: number; brain: number; total: number };
-    limits?: StorageLimits;
-  };
-}
-
-export interface StorageLimits {
-  totalLimitGiB?: number;
-  warnAtPercent?: number;
-}
-
-export interface TokenRecord {
-  id: string;
-  name: string;
-  createdAt: string;
-  lastUsed?: string;
-  expiresAt?: string;
-  spaces?: string[];
-  admin: boolean;
-  readOnly?: boolean;
-  schemaLibrary?: boolean;
-}
-
-export interface Memory {
-  _id: string;
-  fact: string;
-  type?: string;
-  tags?: string[];
-  entityIds?: string[];
-  description?: string;
-  properties?: Record<string, string | number | boolean>;
-  createdAt: string;
-  seq: number;
-  author?: { instanceId: string };
-}
-
-export interface Entity {
-  _id: string;
-  name: string;
-  type?: string;
-  tags?: string[];
-  description?: string;
-  properties?: Record<string, string | number | boolean>;
-  createdAt: string;
-}
-
-export interface Edge {
-  _id: string;
-  from: string;
-  fromName?: string;
-  to: string;
-  toName?: string;
-  label: string;
-  type?: string;
-  weight?: number;
-  tags?: string[];
-  description?: string;
-  properties?: Record<string, string | number | boolean>;
-  createdAt: string;
-}
-
-export type ChronoType = 'event' | 'deadline' | 'plan' | 'prediction' | 'milestone';
-/** @deprecated Use ChronoType */
-export type ChronoKind = ChronoType;
-export type ChronoStatus = 'upcoming' | 'active' | 'completed' | 'overdue' | 'cancelled';
-
-export interface ChronoEntry {
-  _id: string;
-  spaceId: string;
-  title: string;
-  description?: string;
-  type: ChronoType;
-  startsAt: string;
-  endsAt?: string;
-  status: ChronoStatus;
-  confidence?: number;
-  tags: string[];
-  entityIds: string[];
-  memoryIds: string[];
-  recurrence?: { freq: string; interval?: number; until?: string };
-  author: { instanceId: string; instanceLabel: string };
-  createdAt: string;
-  updatedAt: string;
-  seq: number;
-}
-
-export interface SpaceStats {
-  spaceId: string;
-  memories: number;
-  entities: number;
-  edges: number;
-  chrono: number;
-  files: number;
-  needsReindex?: boolean;
-}
-
-export type QueryCollection = 'memories' | 'entities' | 'edges' | 'chrono' | 'files';
-
-export interface QueryResult {
-  results: Record<string, unknown>[];
-  collection: QueryCollection;
-  count: number;
-}
-
-export type WipeCollectionType = 'memories' | 'entities' | 'edges' | 'chrono' | 'files';
-
-export type RecallKnowledgeType = 'memory' | 'entity' | 'edge' | 'chrono' | 'file';
-
-export interface RecallResult {
-  type: RecallKnowledgeType;
-  score?: number;
-  [key: string]: unknown;
-}
-
-export interface RecallResponse {
-  results: RecallResult[];
-  count: number;
-}
-
-export interface TraverseNode {
-  _id: string;
-  name: string;
-  type: string;
-  depth: number;
-  description?: string;
-  tags?: string[];
-}
-
-export interface TraverseEdge {
-  _id: string;
-  from: string;
-  to: string;
-  label: string;
-}
-
-export interface TraverseResult {
-  nodes: TraverseNode[];
-  edges: TraverseEdge[];
-  truncated: boolean;
-}
-
-export interface WipeResult {
-  memories: number;
-  entities: number;
-  edges: number;
-  chrono: number;
-  files: number;
-}
-
-export interface FileEntry {
-  name: string;
-  size: number;
-  isFile: boolean;
-  isDirectory: boolean;
-  modified: string;
-}
-
-export interface FileMeta {
-  _id: string;
-  spaceId: string;
-  path: string;
-  description?: string;
-  tags: string[];
-  entityIds?: string[];
-  chronoIds?: string[];
-  memoryIds?: string[];
-  properties?: Record<string, string | number | boolean>;
-  sizeBytes: number;
-  createdAt: string;
-  updatedAt: string;
-  /** Async embedding lifecycle status (text documents and media files).
-   *  "pending" → queued; "processing" → being embedded; "complete" → done;
-   *  "partial" → stored but some chunks failed to embed (retry-eligible);
-   *  "failed" → all retries exhausted; "skipped" / "disabled" → media-only states. */
-  embeddingStatus?: 'pending' | 'processing' | 'complete' | 'partial' | 'failed' | 'skipped' | 'disabled';
-  /** Error message when embeddingStatus is "failed". */
-  mediaJobError?: string;
-  chunkCount?: number;
-  /** Set when the file was deleted but its metadata was retained (softDeleteFileMeta):
-   *  ISO8601 deletion timestamp. Such records show a "deleted" badge and can be purged. */
-  deletedAt?: string;
-}
-
-export interface UploadProgress {
-  percent: number;
-  done: boolean;
-}
-
-export interface Network {
-  id: string;
-  label: string;
-  type: 'closed' | 'democratic' | 'club' | 'braintree' | 'pubsub';
-  spaces: string[];
-  spaceMap?: Record<string, string>;
-  members: NetworkMember[];
-  votingDeadlineHours?: number;
-  syncSchedule?: string;
-  merkle?: boolean;
-}
-
-export interface NetworkMember {
-  instanceId: string;
-  label: string;
-  endpoint: string;
-  syncDirection?: 'both' | 'push' | 'pull';
-}
-
-export interface InviteBundle {
-  handshakeId: string;
-  inviteUrl: string;
-  rsaPublicKeyPem: string;
-  networkId: string;
-  expiresAt: string;
-  spaces?: string[];
-}
-
-export interface VoteRound {
-  id: string;
-  networkId: string;
-  type: string;
-  subject: string;
-  openedAt: string;
-  deadline: string;
-  status: 'open' | 'passed' | 'failed';
-  votes: { instanceId: string; vote: 'yes' | 'veto'; }[];
-}
-
-export interface ConflictRecord {
-  id: string;
-  spaceId: string;
-  originalPath: string;
-  conflictPath: string;
-  detectedAt: string;
-  peerInstanceId: string;
-  peerInstanceLabel: string;
-}
-
-export interface DupeActionRule {
-  minScore: number;
-  action: 'flag' | 'automerge' | 'notify';
-  types?: string[];
-  webhookUrl?: string;
-}
-
-export interface DuplicateRecord {
-  id: string;
-  spaceId: string;
-  type: string;
-  aId: string;
-  aSummary: string;
-  bId: string;
-  bSummary: string;
-  score: number;
-  status: 'open' | 'dismissed' | 'resolved';
-  resolution?: 'merged' | 'notified';
-  detectedAt: string;
-  updatedAt: string;
-}
-
-export interface SyncHistoryRecord {
-  _id: string;
-  networkId: string;
-  triggeredAt: string;
-  completedAt: string;
-  status: 'success' | 'partial' | 'failed';
-  pulled: { memories: number; entities: number; edges: number; files: number };
-  pushed: { memories: number; entities: number; edges: number; files: number };
-  errors?: string[];
-}
-
-export interface AboutInfo {
-  instanceId: string;
-  instanceLabel: string;
-  version: string;
-  uptime: string;
-  mongoVersion: string;
-  diskInfo: { total: number; used: number; available: number };
-  publicUrl?: string;
-}
-
-export interface LocalAgentStatus {
-  configured: boolean;
-  reachable: boolean;
-  canExecute: boolean;
-  message?: string;
-}
-
-export interface LocalAgentEnableNetworksResult {
-  ok: boolean;
-  publicUrl?: string;
-  message?: string;
-  steps?: string[];
-}
-
-export interface LocalAgentBootstrapResult {
-  ok: boolean;
-  message?: string;
-}
-
-// ── Audit log types ──────────────────────────────────────────────────────────
-
-export interface AuditLogEntry {
-  _id: string;
-  timestamp: string;
-  tokenId: string | null;
-  tokenLabel: string | null;
-  authMethod: 'pat' | 'oidc' | null;
-  oidcSubject: string | null;
-  ip: string;
-  method: string;
-  path: string;
-  spaceId: string | null;
-  operation: string;
-  status: number;
-  entryId: string | null;
-  durationMs: number;
-}
-
-export interface AuditLogResponse {
-  entries: AuditLogEntry[];
-  total: number;
-  hasMore: boolean;
-  retentionDays?: number;
-}
-
-export interface AuditLogParams {
-  after?: string;
-  before?: string;
-  tokenId?: string;
-  oidcSubject?: string;
-  spaceId?: string;
-  operation?: string;
-  status?: number;
-  ip?: string;
-  limit?: number;
-  offset?: number;
-}
-
-// ── Backup config ─────────────────────────────────────────────────────────────
-
-export interface BackupConfigData {
-  schedule?: string;
-  retention?: { keepLocal?: number };
-  offsite?: { destPath: string; retention?: { keepCount?: number } };
-}
-
-// ── API service ───────────────────────────────────────────────────────────────
-
-@Injectable({ providedIn: 'root' })
-export class ApiService {
-  private http = inject(HttpClient);
-
-  // ── Auth ──────────────────────────────────────────────────────────────────
-
-  /** Verify the supplied PAT is valid and return its metadata */
-  verifyToken(): Observable<{ id: string; name: string; spaces?: string[] }> {
-    return this.http.get<any>('/api/tokens/me');
-  }
-
-  // ── Setup ─────────────────────────────────────────────────────────────────
-
-  getSetupStatus(): Observable<{ configured: boolean }> {
-    return this.http.get<{ configured: boolean }>('/api/setup/status');
-  }
-
-  completeSetup(body: {
-    code: string;
-    label: string;
-    settingsPassword: string;
-  }): Observable<{ plaintext: string }> {
-    return this.http.post<{ plaintext: string }>('/api/setup', body);
-  }
-
-  // ── Schema Library ─────────────────────────────────────────────────────────
-
-  listSchemaLibrary(): Observable<{ entries: SchemaLibraryEntry[] }> {
-    return this.http.get<{ entries: SchemaLibraryEntry[] }>('/api/schema-library');
-  }
-
-  getSchemaLibraryEntry(name: string): Observable<{ entry: SchemaLibraryEntry }> {
-    return this.http.get<{ entry: SchemaLibraryEntry }>(`/api/schema-library/${encodeURIComponent(name)}`);
-  }
-
-  createSchemaLibraryEntry(body: Omit<SchemaLibraryEntry, 'createdAt' | 'updatedAt'>): Observable<{ entry: SchemaLibraryEntry }> {
-    return this.http.post<{ entry: SchemaLibraryEntry }>('/api/schema-library', body);
-  }
-
-  upsertSchemaLibraryEntry(name: string, body: Omit<SchemaLibraryEntry, 'name' | 'createdAt' | 'updatedAt'>): Observable<{ entry: SchemaLibraryEntry }> {
-    return this.http.put<{ entry: SchemaLibraryEntry }>(`/api/schema-library/${encodeURIComponent(name)}`, body);
-  }
-
-  deleteSchemaLibraryEntry(name: string): Observable<void> {
-    return this.http.delete<void>(`/api/schema-library/${encodeURIComponent(name)}`);
-  }
-
-  getSchemaLibraryUsages(name: string): Observable<{ usages: { spaceId: string; spaceLabel: string; knowledgeType: string; typeName: string }[] }> {
-    return this.http.get<{ usages: { spaceId: string; spaceLabel: string; knowledgeType: string; typeName: string }[] }>(`/api/schema-library/${encodeURIComponent(name)}/usages`);
-  }
-
-  publishSchemaLibraryEntry(name: string, published: boolean): Observable<{ entry: SchemaLibraryEntry }> {
-    return this.http.patch<{ entry: SchemaLibraryEntry }>(`/api/schema-library/${encodeURIComponent(name)}/publish`, { published });
-  }
-
-  getPublicSchemaLibrary(): Observable<{ entries: ForeignCatalogEntry[] }> {
-    return this.http.get<{ entries: ForeignCatalogEntry[] }>('/api/schema-library/public');
-  }
-
-  /** List all distinct schema group names and their entry counts. */
-  listSchemaLibraryGroups(): Observable<{ groups: { name: string; count: number }[] }> {
-    return this.http.get<{ groups: { name: string; count: number }[] }>('/api/schema-library/groups');
-  }
-
-  /** Export a space's full typeSchemas into the library as a named group. */
-  exportSpaceSchemaToLibrary(body: { spaceId: string; groupName: string; namePrefix?: string }): Observable<{ created: number; updated: number; entries: SchemaLibraryEntry[] }> {
-    return this.http.post<{ created: number; updated: number; entries: SchemaLibraryEntry[] }>('/api/schema-library/export-space', body);
-  }
-
-  /** Apply all library entries belonging to a group to a space as $ref links. */
-  applyGroupToSpace(group: string, spaceId: string): Observable<{ applied: { knowledgeType: string; typeName: string; entryName: string }[]; count: number }> {
-    return this.http.post<{ applied: { knowledgeType: string; typeName: string; entryName: string }[]; count: number }>(`/api/schema-library/groups/${encodeURIComponent(group)}/apply`, { spaceId });
-  }
-
-  // ── Schema catalogs ────────────────────────────────────────────────────────
-
-  listSchemaCatalogs(): Observable<{ catalogs: SchemaCatalog[] }> {
-    return this.http.get<{ catalogs: SchemaCatalog[] }>('/api/schema-library/catalogs');
-  }
-
-  addSchemaCatalog(body: { name: string; url: string; description?: string; accessToken?: string }): Observable<{ catalog: SchemaCatalog }> {
-    return this.http.post<{ catalog: SchemaCatalog }>('/api/schema-library/catalogs', body);
-  }
-
-  deleteSchemaCatalog(name: string): Observable<void> {
-    return this.http.delete<void>(`/api/schema-library/catalogs/${encodeURIComponent(name)}`);
-  }
-
-  browseCatalog(catalogName: string): Observable<{ catalog: string; entries: ForeignCatalogEntry[] }> {
-    return this.http.get<{ catalog: string; entries: ForeignCatalogEntry[] }>(`/api/schema-library/catalogs/${encodeURIComponent(catalogName)}/entries`);
-  }
-
-  getCatalogEntry(catalogName: string, entryName: string): Observable<{ catalog: string; entry: ForeignCatalogEntry }> {
-    return this.http.get<{ catalog: string; entry: ForeignCatalogEntry }>(`/api/schema-library/catalogs/${encodeURIComponent(catalogName)}/entries/${encodeURIComponent(entryName)}`);
-  }
-
-  // ── Spaces ────────────────────────────────────────────────────────────────
-
-  listSpaces(): Observable<SpacesResponse> {
-    return this.http.get<SpacesResponse>('/api/spaces');
-  }
-
-  createSpace(body: { label: string; id?: string; maxGiB?: number; description?: string; proxyFor?: string[]; meta?: Partial<SpaceMeta> }): Observable<{ space: Space }> {
-    return this.http.post<{ space: Space }>('/api/spaces', body);
-  }
-
-  updateSpace(id: string, body: { label?: string; description?: string; maxGiB?: number | null; meta?: Partial<SpaceMeta>; dupeRules?: DupeActionRule[]; dupeMergeSurvivor?: 'older' | 'newer'; dupeRulesOnInsert?: boolean }): Observable<{ space: Space }> {
-    return this.http.patch<{ space: Space }>(`/api/spaces/${id}`, body);
-  }
-
-  reorderSpaces(ids: string[]): Observable<{ spaces: Space[] }> {
-    return this.http.post<{ spaces: Space[] }>('/api/spaces/reorder', { ids });
-  }
-
-  getSpaceMeta(id: string): Observable<SpaceMetaResponse> {
-    return this.http.get<SpaceMetaResponse>(`/api/spaces/${id}/meta`);
-  }
-
-  /** GET a single type definition from the space's typeSchemas. */
-  getTypeSchema(spaceId: string, knowledgeType: KnowledgeType, typeName: string): Observable<{ knowledgeType: KnowledgeType; typeName: string; schema: TypeSchema }> {
-    return this.http.get<{ knowledgeType: KnowledgeType; typeName: string; schema: TypeSchema }>(
-      `/api/spaces/${spaceId}/meta/typeSchemas/${knowledgeType}/${typeName}`,
-    );
-  }
-
-  /** PUT (upsert) a single type definition into the space's typeSchemas. */
-  upsertTypeSchema(spaceId: string, knowledgeType: KnowledgeType, typeName: string, schema: TypeSchema): Observable<{ knowledgeType: KnowledgeType; typeName: string; schema: TypeSchema }> {
-    return this.http.put<{ knowledgeType: KnowledgeType; typeName: string; schema: TypeSchema }>(
-      `/api/spaces/${spaceId}/meta/typeSchemas/${knowledgeType}/${typeName}`,
-      schema,
-    );
-  }
-
-  /** DELETE a single type definition from the space's typeSchemas. */
-  deleteTypeSchema(spaceId: string, knowledgeType: KnowledgeType, typeName: string): Observable<void> {
-    return this.http.delete<void>(`/api/spaces/${spaceId}/meta/typeSchemas/${knowledgeType}/${typeName}`);
-  }
-
-  deleteSpace(id: string): Observable<void> {
-    return this.http.delete<void>(`/api/spaces/${id}`, { body: { confirm: true } });
-  }
-
-  renameSpace(oldId: string, newId: string): Observable<{ space: Space }> {
-    return this.http.patch<{ space: Space }>(`/api/spaces/${oldId}/rename`, { newId });
-  }
-
-  // ── Tokens ────────────────────────────────────────────────────────────────
-
-  getMe(): Observable<TokenRecord> {
-    return this.http.get<TokenRecord>('/api/tokens/me');
-  }
-
-  listTokens(): Observable<{ tokens: TokenRecord[] }> {
-    return this.http.get<{ tokens: TokenRecord[] }>('/api/tokens');
-  }
-
-  createToken(body: { name: string; expiresAt?: string; spaces?: string[]; admin?: boolean; readOnly?: boolean; schemaLibrary?: boolean }): Observable<{ token: TokenRecord; plaintext: string }> {
-    return this.http.post<{ token: TokenRecord; plaintext: string }>('/api/tokens', body);
-  }
-
-  regenerateToken(id: string): Observable<{ plaintext: string }> {
-    return this.http.post<{ plaintext: string }>(`/api/tokens/${id}/regenerate`, {});
-  }
-
-  revokeToken(id: string): Observable<void> {
-    return this.http.delete<void>(`/api/tokens/${id}`);
-  }
-
-  wipeSpace(spaceId: string, types?: WipeCollectionType[]): Observable<{ deleted: WipeResult }> {
-    const body: { types?: WipeCollectionType[] } = types && types.length > 0 ? { types } : {};
-    return this.http.post<{ deleted: WipeResult }>(`/api/admin/spaces/${spaceId}/wipe`, body);
-  }
-
-  // ── MFA ───────────────────────────────────────────────────────────────────
-
-  getMfaStatus(): Observable<{ enabled: boolean }> {
-    return this.http.get<{ enabled: boolean }>('/api/mfa/status');
-  }
-
-  setupMfa(): Observable<{ secret: string; otpauth: string }> {
-    return this.http.post<{ secret: string; otpauth: string }>('/api/mfa/setup', {});
-  }
-
-  verifyMfaCode(code: string): Observable<{ valid: boolean }> {
-    return this.http.post<{ valid: boolean }>('/api/mfa/verify', { code });
-  }
-
-  disableMfa(): Observable<void> {
-    return this.http.delete<void>('/api/mfa');
-  }
-
-  getSpaceStats(spaceId: string): Observable<SpaceStats> {
-    return this.http.get<SpaceStats>(`/api/brain/spaces/${spaceId}/stats`);
-  }
-
-  getReindexStatus(spaceId: string): Observable<{ spaceId: string; needsReindex: boolean }> {
-    return this.http.get<{ spaceId: string; needsReindex: boolean }>(`/api/brain/spaces/${spaceId}/reindex-status`);
-  }
-
-  reindex(spaceId: string): Observable<Record<string, number>> {
-    return this.http.post<Record<string, number>>(`/api/brain/spaces/${spaceId}/reindex`, {});
-  }
-
-  queryBrain(
-    spaceId: string,
-    body: {
-      collection: QueryCollection;
-      filter?: Record<string, unknown>;
-      projection?: Record<string, unknown>;
-      limit?: number;
-      maxTimeMS?: number;
-    },
-  ): Observable<QueryResult> {
-    return this.http.post<QueryResult>(`/api/brain/spaces/${spaceId}/query`, body);
-  }
-
-  recallBrain(
-    spaceId: string,
-    body: {
-      query: string;
-      topK?: number;
-      types?: RecallKnowledgeType[];
-      minScore?: number;
-      /** Structured filter (same expression grammar as the query filter). */
-      filter?: Record<string, unknown>;
-      /** Restrict to records carrying these tags. */
-      tags?: string[];
-      /** Guarantee at least N hits per knowledge type, e.g. { entity: 2 }. */
-      minPerType?: Partial<Record<RecallKnowledgeType, number>>;
-    },
-  ): Observable<RecallResponse> {
-    return this.http.post<RecallResponse>(`/api/brain/spaces/${spaceId}/recall`, body);
-  }
-
-  // ── Brain — memories ──────────────────────────────────────────────────────
-
-  listMemories(spaceId: string, limit = 20, skip = 0, filters?: { tag?: string; entity?: string; type?: string }): Observable<{ memories: Memory[]; limit: number; skip: number }> {
-    let params = new HttpParams().set('limit', limit).set('skip', skip);
-    if (filters?.tag) params = params.set('tag', filters.tag);
-    if (filters?.entity) params = params.set('entity', filters.entity);
-    if (filters?.type) params = params.set('type', filters.type);
-    return this.http.get<any>(`/api/brain/spaces/${spaceId}/memories`, { params });
-  }
-
-  deleteMemory(spaceId: string, id: string): Observable<void> {
-    return this.http.delete<void>(`/api/brain/spaces/${spaceId}/memories/${id}`);
-  }
-
-  createMemory(spaceId: string, body: { fact: string; tags?: string[]; entityIds?: string[]; description?: string; properties?: Record<string, string | number | boolean> }): Observable<Memory> {
-    return this.http.post<Memory>(`/api/brain/spaces/${spaceId}/memories`, body);
-  }
-
-  updateMemory(spaceId: string, id: string, body: Partial<{ fact: string; tags: string[]; entityIds: string[]; description: string; properties: Record<string, string | number | boolean>; deleteFields: string[] }>): Observable<Memory> {
-    return this.http.patch<Memory>(`/api/brain/spaces/${spaceId}/memories/${id}`, body);
-  }
-
-  wipeMemories(spaceId: string): Observable<{ deleted: number }> {
-    return this.http.delete<{ deleted: number }>(`/api/brain/spaces/${spaceId}/memories`, {
-      body: { confirm: true },
-    });
-  }
-
-  // ── Brain — entities ──────────────────────────────────────────────────────
-
-  listEntities(spaceId: string, limit = 50, skip = 0, filters?: { search?: string; type?: string; tag?: string }): Observable<{ entities: Entity[] }> {
-    let params = new HttpParams().set('limit', limit).set('skip', skip);
-    if (filters?.search) params = params.set('name', filters.search);
-    if (filters?.type) params = params.set('type', filters.type);
-    if (filters?.tag) params = params.set('tag', filters.tag);
-    return this.http.get<any>(`/api/brain/spaces/${spaceId}/entities`, { params });
-  }
-
-  deleteEntity(spaceId: string, id: string): Observable<void> {
-    return this.http.delete<void>(`/api/brain/spaces/${spaceId}/entities/${id}`);
-  }
-
-  createEntity(spaceId: string, body: { name: string; type?: string; tags?: string[]; description?: string; properties?: Record<string, string | number | boolean> }): Observable<Entity> {
-    return this.http.post<Entity>(`/api/brain/spaces/${spaceId}/entities`, body);
-  }
-
-  updateEntity(spaceId: string, id: string, body: Partial<{ name: string; type: string; description: string; tags: string[]; properties: Record<string, string | number | boolean>; deleteFields: string[] }>): Observable<Entity> {
-    return this.http.patch<Entity>(`/api/brain/spaces/${spaceId}/entities/${id}`, body);
-  }
-
-  // ── Brain — edges ─────────────────────────────────────────────────────────
-
-  listEdges(spaceId: string, limit = 50, skip = 0, filters?: { type?: string; tag?: string }): Observable<{ edges: Edge[] }> {
-    let params = new HttpParams().set('limit', limit).set('skip', skip);
-    if (filters?.type) params = params.set('type', filters.type);
-    if (filters?.tag) params = params.set('tag', filters.tag);
-    return this.http.get<any>(`/api/brain/spaces/${spaceId}/edges`, { params });
-  }
-
-  deleteEdge(spaceId: string, id: string): Observable<void> {
-    return this.http.delete<void>(`/api/brain/spaces/${spaceId}/edges/${id}`);
-  }
-
-  createEdge(spaceId: string, body: { from: string; to: string; label: string; weight?: number; type?: string; tags?: string[]; description?: string; properties?: Record<string, string | number | boolean> }): Observable<Edge> {
-    return this.http.post<Edge>(`/api/brain/spaces/${spaceId}/edges`, body);
-  }
-
-  updateEdge(spaceId: string, id: string, body: Partial<{ label: string; description: string; tags: string[]; properties: Record<string, string | number | boolean>; weight: number; type: string; deleteFields: string[] }>): Observable<Edge> {
-    return this.http.patch<Edge>(`/api/brain/spaces/${spaceId}/edges/${id}`, body);
-  }
-
-  // ── Brain — graph traverse ────────────────────────────────────────────
-
-  searchEntitiesByName(spaceId: string, name: string): Observable<{ entities: Entity[] }> {
-    const params = new HttpParams().set('name', name);
-    return this.http.get<{ entities: Entity[] }>(`/api/brain/spaces/${spaceId}/entities/by-name`, { params });
-  }
-
-  getEntitiesByIds(spaceId: string, ids: string[]): Observable<{ entities: Entity[] }> {
-    if (!ids.length) return new Observable(o => { o.next({ entities: [] }); o.complete(); });
-    const params = new HttpParams().set('ids', ids.join(','));
-    return this.http.get<{ entities: Entity[] }>(`/api/brain/spaces/${spaceId}/entities/by-ids`, { params });
-  }
-
-  getEntity(spaceId: string, id: string): Observable<Entity> {
-    return this.http.get<Entity>(`/api/brain/spaces/${spaceId}/entities/${id}`);
-  }
-
-  getEdge(spaceId: string, id: string): Observable<Edge> {
-    return this.http.get<Edge>(`/api/brain/spaces/${spaceId}/edges/${id}`);
-  }
-
-  getMemory(spaceId: string, id: string): Observable<Memory> {
-    return this.http.get<Memory>(`/api/brain/spaces/${spaceId}/memories/${id}`);
-  }
-
-  getChrono(spaceId: string, id: string): Observable<ChronoEntry> {
-    return this.http.get<ChronoEntry>(`/api/brain/spaces/${spaceId}/chrono/${id}`);
-  }
-
-  traverseGraph(spaceId: string, body: { startId: string; direction?: 'outbound' | 'inbound' | 'both'; maxDepth?: number; limit?: number }): Observable<TraverseResult> {
-    return this.http.post<TraverseResult>(`/api/brain/spaces/${spaceId}/traverse`, body);
-  }
-
-  // ── Brain — chrono ──────────────────────────────────────────────────────
-
-  listChrono(spaceId: string, limit = 50, skip = 0, filters?: { tags?: string; tagsAny?: string; tag?: string; type?: string; status?: string; after?: string; before?: string; search?: string }): Observable<{ chrono: ChronoEntry[] }> {
-    let params = new HttpParams().set('limit', limit).set('skip', skip);
-    if (filters?.tags) params = params.set('tags', filters.tags);
-    if (filters?.tagsAny) params = params.set('tagsAny', filters.tagsAny);
-    if (filters?.tag) params = params.set('tag', filters.tag);
-    // The chrono record "kind" (event/deadline/…) is filtered via the `type` query
-    // param server-side; the old `kind` param was silently ignored.
-    if (filters?.type) params = params.set('type', filters.type);
-    if (filters?.status) params = params.set('status', filters.status);
-    if (filters?.after) params = params.set('after', filters.after);
-    if (filters?.before) params = params.set('before', filters.before);
-    if (filters?.search) params = params.set('search', filters.search);
-    return this.http.get<any>(`/api/brain/spaces/${spaceId}/chrono`, { params });
-  }
-
-  createChrono(spaceId: string, body: { title: string; type: ChronoType; startsAt: string; endsAt?: string; status?: ChronoStatus; confidence?: number; tags?: string[]; entityIds?: string[]; memoryIds?: string[]; description?: string }): Observable<ChronoEntry> {
-    return this.http.post<ChronoEntry>(`/api/brain/spaces/${spaceId}/chrono`, body);
-  }
-
-  updateChrono(spaceId: string, id: string, body: Partial<{ title: string; type: ChronoType; startsAt: string; endsAt: string; status: ChronoStatus; confidence: number; tags: string[]; entityIds: string[]; memoryIds: string[]; description: string }>): Observable<ChronoEntry> {
-    return this.http.post<ChronoEntry>(`/api/brain/spaces/${spaceId}/chrono/${id}`, body);
-  }
-
-  deleteChrono(spaceId: string, id: string): Observable<void> {
-    return this.http.delete<void>(`/api/brain/spaces/${spaceId}/chrono/${id}`);
-  }
-
-  // ── Files ─────────────────────────────────────────────────────────────────
-
-  listFiles(spaceId: string, path = '/'): Observable<{ entries: FileEntry[] }> {
-    const params = new HttpParams().set('path', path);
-    return this.http.get<any>(`/api/files/${spaceId}`, { params }).pipe(
-      map(res => ({
-        entries: (res.entries ?? []).map((e: any) => ({
-          name: e.name,
-          size: e.size ?? 0,
-          isFile: e.type === 'file',
-          isDirectory: e.type === 'dir',
-          modified: e.modifiedAt ?? '',
-        } as FileEntry)),
-      })),
-    );
-  }
-
-  deleteFile(spaceId: string, path: string): Observable<void> {
-    const params = new HttpParams().set('path', path);
-    return this.http.delete<void>(`/api/files/${spaceId}`, { params, body: { confirm: true } });
-  }
-
-  createDir(spaceId: string, path: string): Observable<void> {
-    const params = new HttpParams().set('path', path);
-    return this.http.post<void>(`/api/files/${spaceId}/mkdir`, null, { params });
-  }
-
-  moveFile(spaceId: string, from: string, to: string): Observable<void> {
-    const params = new HttpParams().set('path', from);
-    return this.http.patch<void>(`/api/files/${spaceId}`, { destination: to }, { params });
-  }
-
-  /**
-   * Upload a file with automatic chunking for files > 10 MB.
-   * Emits progress events ({ percent, done }) for UI updates.
-   * Retries each chunk up to 3 times on transient failure.
-   *
-   * The returned observable is **cold**: no work runs and no progress event is
-   * emitted until the caller subscribes, so a late subscriber can never miss the
-   * initial `{ percent: 0 }` (with the old hot Subject the upload started before
-   * `return`, and any event emitted synchronously was lost). Unsubscribing tears
-   * the upload down — it flips a `cancelled` flag that halts the chunk loop and
-   * unsubscribes the in-flight request (HttpClient aborts the underlying XHR on
-   * teardown). That is exactly how the UI cancels an upload mid-flight.
-   */
-  uploadFileChunked(spaceId: string, dirPath: string, file: File): Observable<UploadProgress> {
-    const CHUNK_THRESHOLD = 10 * 1024 * 1024; // 10 MB
-    const CHUNK_SIZE = 5 * 1024 * 1024; // 5 MB
-    const MAX_RETRIES = 3;
-    const filePath = dirPath.endsWith('/') ? `${dirPath}${file.name}` : `${dirPath}/${file.name}`;
-
-    return new Observable<UploadProgress>(subscriber => {
-      let cancelled = false;
-      let httpSub: Subscription | null = null;
-      // A rejected arrayBuffer() (e.g. the file moved/was revoked) must surface,
-      // but never after the caller unsubscribed.
-      const fail = (err: unknown): void => { if (!cancelled) subscriber.error(err); };
-
-      if (file.size <= CHUNK_THRESHOLD) {
-        // Small file: single upload.
-        subscriber.next({ percent: 0, done: false });
-        file.arrayBuffer().then(ab => {
-          if (cancelled) return;
-          const headers = new HttpHeaders({ 'Content-Type': 'application/octet-stream' });
-          const params = new HttpParams().set('path', filePath);
-          httpSub = this.http.post<void>(`/api/files/${spaceId}`, ab, { headers, params }).subscribe({
-            next: () => {
-              subscriber.next({ percent: 100, done: true });
-              subscriber.complete();
-            },
-            error: fail,
-          });
-        }).catch(fail);
-      } else {
-        // Chunked upload.
-        const total = file.size;
-        let offset = 0;
-
-        const sendNextChunk = (): void => {
-          if (cancelled || offset >= total) return;
-          const end = Math.min(offset + CHUNK_SIZE, total);
-          const slice = file.slice(offset, end);
-          const start = offset;
-          const byteEnd = end - 1;
-
-          slice.arrayBuffer().then(ab => {
-            if (cancelled) return;
-            const sendChunk = (retriesLeft: number): void => {
-              if (cancelled) return;
-              const headers = new HttpHeaders({
-                'Content-Type': 'application/octet-stream',
-                'Content-Range': `bytes ${start}-${byteEnd}/${total}`,
-              });
-              const params = new HttpParams().set('path', filePath);
-              httpSub = this.http.post<any>(`/api/files/${spaceId}`, ab, { headers, params }).subscribe({
-                next: () => {
-                  offset = end;
-                  const percent = Math.round((offset / total) * 100);
-                  if (offset >= total) {
-                    subscriber.next({ percent: 100, done: true });
-                    subscriber.complete();
-                  } else {
-                    subscriber.next({ percent, done: false });
-                    sendNextChunk();
-                  }
-                },
-                error: err => {
-                  if (cancelled) return;
-                  if (retriesLeft > 0) {
-                    sendChunk(retriesLeft - 1);
-                  } else {
-                    subscriber.error(err);
-                  }
-                },
-              });
-            };
-            sendChunk(MAX_RETRIES);
-          }).catch(fail);
-        };
-
-        subscriber.next({ percent: 0, done: false });
-        sendNextChunk();
-      }
-
-      // Teardown — cancel: stop the loop and abort any in-flight chunk request.
-      return () => {
-        cancelled = true;
-        httpSub?.unsubscribe();
-      };
-    });
-  }
-
-  getFileDownloadUrl(spaceId: string, path: string): string {
-    return `/api/files/${spaceId}?path=${encodeURIComponent(path)}`;
-  }
-
-  // ── File metadata (brain) ─────────────────────────────────────────────────
-
-  listFileMeta(spaceId: string, limit = 50, skip = 0, search?: string): Observable<{ files: FileMeta[]; limit: number; skip: number }> {
-    let params = new HttpParams().set('limit', limit).set('skip', skip);
-    if (search) params = params.set('path', search);
-    return this.http.get<{ files: FileMeta[]; limit: number; skip: number }>(`/api/brain/spaces/${spaceId}/files`, { params });
-  }
-
-  updateFileMeta(spaceId: string, path: string, body: Partial<{ description: string; tags: string[]; entityIds: string[]; chronoIds: string[]; memoryIds: string[]; properties: Record<string, string | number | boolean> }>): Observable<FileMeta> {
-    const params = new HttpParams().set('path', path);
-    return this.http.patch<FileMeta>(`/api/brain/spaces/${spaceId}/files`, body, { params });
-  }
-
-  /** Re-queue embedding for a file whose embedding failed or is only partial. */
-  retryEmbedding(spaceId: string, path: string): Observable<{ queued: boolean }> {
-    const params = new HttpParams().set('path', path);
-    return this.http.post<{ queued: boolean }>(`/api/files/${spaceId}/retry_embedding`, {}, { params });
-  }
-
-  deleteFileMeta(spaceId: string, path: string): Observable<void> {
-    const params = new HttpParams().set('path', path);
-    return this.http.delete<void>(`/api/brain/spaces/${spaceId}/files`, { params });
-  }
-
-  // ── File conflicts ────────────────────────────────────────────────────────
-
-  listConflicts(): Observable<{ conflicts: ConflictRecord[] }> {
-    return this.http.get<any>('/api/conflicts');
-  }
-
-  resolveConflict(id: string, action: string = 'keep-local', opts?: { rename?: string; targetSpaceId?: string }): Observable<{ status: string }> {
-    return this.http.post<{ status: string }>(`/api/conflicts/${id}/resolve`, { action, ...opts });
-  }
-
-  bulkResolveConflicts(ids: string[], action: string, opts?: { rename?: string; targetSpaceId?: string }): Observable<{ resolved: number; failed: { id: string; error: string }[] }> {
-    return this.http.post<any>('/api/conflicts/bulk-resolve', { ids, action, ...opts });
-  }
-
-  dismissConflict(id: string): Observable<void> {
-    return this.http.delete<void>(`/api/conflicts/${id}`);
-  }
-
-  // ── Duplicate candidates ──────────────────────────────────────────────────
-
-  listDuplicates(status: 'open' | 'dismissed' | 'all' = 'open', space?: string): Observable<{ duplicates: DuplicateRecord[] }> {
-    const params = new URLSearchParams({ status });
-    if (space) params.set('space', space);
-    return this.http.get<{ duplicates: DuplicateRecord[] }>(`/api/duplicates?${params.toString()}`);
-  }
-
-  dismissDuplicate(id: string): Observable<{ status: string }> {
-    return this.http.post<{ status: string }>(`/api/duplicates/${encodeURIComponent(id)}/dismiss`, {});
-  }
-
-  mergeDuplicate(id: string): Observable<{ status: string; survivorId?: string }> {
-    return this.http.post<{ status: string; survivorId?: string }>(`/api/duplicates/${encodeURIComponent(id)}/merge`, {});
-  }
-
-  scanDuplicates(space?: string): Observable<{ scannedSpaces: number; scanned: number; pairs: number }> {
-    return this.http.post<{ scannedSpaces: number; scanned: number; pairs: number }>(`/api/duplicates/scan${space ? `?space=${encodeURIComponent(space)}` : ''}`, {});
-  }
-
-  // ── Networks ──────────────────────────────────────────────────────────────
-
-  listNetworks(): Observable<{ networks: Network[] }> {
-    return this.http.get<{ networks: Network[] }>('/api/networks');
-  }
-
-  getNetwork(id: string): Observable<Network> {
-    return this.http.get<Network>(`/api/networks/${id}`);
-  }
-
-  getSyncHistory(networkId: string, limit: number = 20): Observable<{ history: SyncHistoryRecord[] }> {
-    return this.http.get<{ history: SyncHistoryRecord[] }>(`/api/networks/${networkId}/sync-history?limit=${limit}`);
-  }
-
-  createNetwork(body: {
-    label: string;
-    type: string;
-    spaces: string[];
-    votingDeadlineHours?: number;
-    syncSchedule?: string;
-    merkle?: boolean;
-  }): Observable<Network> {
-    return this.http.post<Network>('/api/networks', body);
-  }
-
-  leaveNetwork(id: string): Observable<void> {
-    return this.http.delete<void>(`/api/networks/${id}`, { body: { confirm: true } });
-  }
-
-  generateInvite(networkId: string): Observable<InviteBundle> {
-    return this.http.post<InviteBundle>('/api/invite/generate', { networkId });
-  }
-
-  joinRemote(body: {
-    handshakeId: string;
-    inviteUrl: string;
-    rsaPublicKeyPem: string;
-    networkId: string;
-    myUrl: string;
-    expiresAt?: string;
-    spaceMap?: Record<string, string>;
-  }): Observable<{ status: string; networkId: string; networkLabel: string; networkType: string; spaces: string[]; existingSpaces?: string[]; createdSpaces?: string[]; spaceMap?: Record<string, string>; instanceId?: string; instanceLabel?: string }> {
-    return this.http.post<any>('/api/networks/join-remote', body);
-  }
-
-  removeMember(networkId: string, instanceId: string): Observable<void> {
-    return this.http.delete<void>(`/api/networks/${networkId}/members/${instanceId}`);
-  }
-
-  updateNetworkSchedule(networkId: string, syncSchedule: string): Observable<any> {
-    return this.http.patch<any>(`/api/networks/${networkId}`, { syncSchedule });
-  }
-
-  updateSyncSchedule(networkId: string, memberId: string, schedule: string): Observable<void> {
-    return this.http.patch<void>(`/api/networks/${networkId}/members/${memberId}`, { syncSchedule: schedule });
-  }
-
-  triggerSync(networkId: string): Observable<{ ok: boolean }> {
-    return this.http.post<{ ok: boolean }>(`/api/networks/${networkId}/sync`, {});
-  }
-
-  castVote(networkId: string, roundId: string, vote: 'yes' | 'veto'): Observable<void> {
-    return this.http.post<void>(`/api/networks/${networkId}/votes/${roundId}`, { vote });
-  }
-
-  listVotes(networkId: string): Observable<{ rounds: VoteRound[] }> {
-    return this.http.get<any>(`/api/networks/${networkId}/votes`);
-  }
-
-  // ── About ───────────────────────────────────────────────────────────────
-
-  getAbout(): Observable<AboutInfo> {
-    return this.http.get<AboutInfo>('/api/about');
-  }
-
-  getAboutLogs(lines: number = 200): Observable<{ lines: string[] }> {
-    return this.http.get<{ lines: string[] }>(`/api/about/logs?lines=${lines}`);
-  }
-
-  getLocalAgentStatus(): Observable<LocalAgentStatus> {
-    return this.http.get<LocalAgentStatus>('/api/admin/local-agent/status');
-  }
-
-  bootstrapLocalAgent(body: {
-    os: 'windows' | 'linux';
-  }): Observable<LocalAgentBootstrapResult> {
-    return this.http.post<LocalAgentBootstrapResult>('/api/admin/local-agent/bootstrap', body);
-  }
-
-  executeEnableNetworksViaLocalAgent(body: {
-    hostname: string;
-    os: 'windows' | 'linux';
-    autostart: boolean;
-    overwriteDns: boolean;
-    acknowledgeCriticalChanges: boolean;
-  }): Observable<LocalAgentEnableNetworksResult> {
-    return this.http.post<LocalAgentEnableNetworksResult>('/api/admin/local-agent/enable-networks/execute', body);
-  }
-
-  // ── Audit Log ───────────────────────────────────────────────────────────
-
-  getAuditLog(params: AuditLogParams = {}): Observable<AuditLogResponse> {
-    let p = new HttpParams();
-    if (params.after) p = p.set('after', params.after);
-    if (params.before) p = p.set('before', params.before);
-    if (params.tokenId) p = p.set('tokenId', params.tokenId);
-    if (params.oidcSubject) p = p.set('oidcSubject', params.oidcSubject);
-    if (params.spaceId) p = p.set('spaceId', params.spaceId);
-    if (params.operation) p = p.set('operation', params.operation);
-    if (params.status !== undefined) p = p.set('status', String(params.status));
-    if (params.ip) p = p.set('ip', params.ip);
-    if (params.limit !== undefined) p = p.set('limit', String(params.limit));
-    if (params.offset !== undefined) p = p.set('offset', String(params.offset));
-    return this.http.get<AuditLogResponse>('/api/admin/audit-log', { params: p });
-  }
-
-  // ── Data management ───────────────────────────────────────────────────────
-
-  getDataConfig(): Observable<{ source: 'env' | 'config' | 'default'; mongoUriRedacted: string; migrationEnabled: boolean }> {
-    return this.http.get<{ source: 'env' | 'config' | 'default'; mongoUriRedacted: string; migrationEnabled: boolean }>('/api/admin/data/config');
-  }
-
-  testMongoConnection(uri: string): Observable<{ ok: boolean; error?: string }> {
-    return this.http.post<{ ok: boolean; error?: string }>('/api/admin/data/config/test', { uri });
-  }
-
-  getMaintenanceStatus(): Observable<{ active: boolean }> {
-    return this.http.get<{ active: boolean }>('/api/admin/data/maintenance');
-  }
-
-  setMaintenance(active: boolean): Observable<{ active: boolean }> {
-    return this.http.post<{ active: boolean }>('/api/admin/data/maintenance', { active });
-  }
-
-  triggerBackup(): Observable<{ backup: { id: string; dir: string; manifest: unknown } }> {
-    return this.http.post<{ backup: { id: string; dir: string; manifest: unknown } }>('/api/admin/data/backup', {});
-  }
-
-  listBackups(): Observable<{ backups: Array<{ id: string; createdAt: string; collections: unknown[] }> }> {
-    return this.http.get<{ backups: Array<{ id: string; createdAt: string; collections: unknown[] }> }>('/api/admin/data/backups');
-  }
-
-  restoreBackup(backupId: string): Observable<{ ok: boolean }> {
-    return this.http.post<{ ok: boolean }>('/api/admin/data/restore', { backupId });
-  }
-
-  startMigration(uri: string): Observable<{ ok: boolean; backupDir: string; manifest: unknown }> {
-    return this.http.post<{ ok: boolean; backupDir: string; manifest: unknown }>('/api/admin/data/migrate', { uri });
-  }
-
-  getBackupConfig(): Observable<{ config: BackupConfigData | null; backupsPath?: string }> {
-    return this.http.get<{ config: BackupConfigData | null; backupsPath?: string }>('/api/admin/data/backup-config');
-  }
-
-  saveBackupConfig(config: BackupConfigData): Observable<{ ok: boolean; config: BackupConfigData }> {
-    return this.http.put<{ ok: boolean; config: BackupConfigData }>('/api/admin/data/backup-config', config);
-  }
-
-  browseDirs(dirPath: string): Observable<{ path: string; dirs: string[] }> {
-    return this.http.get<{ path: string; dirs: string[] }>('/api/admin/data/browse-dirs', {
-      params: { path: dirPath },
-    });
-  }
-}
